Make the ui feature build clippy-clean and check it in CI (#164)
* Make the ui feature build clippy-clean and check it in CI

CI ran app clippy only with --no-default-features, so the ui build
(the whole launcher) had accumulated 143 clippy errors, plus a few in
the ui-preview binary and the particle_golden test.

- Apply clippy's machine fixes: let-chains for collapsible ifs, ? for
  early None returns, is_none_or, is_multiple_of, redundant casts and
  struct updates, Default impls for new-only types.
- Rewrite index loops in the CRT backdrop and raw preview fade as
  iterator loops, merge identical if branches, and use as_chunks for
  fixed-size byte chunks (clippy's own rewrite of the mutable case was
  wrong, so those are hand-written).
- Keep device-scan application exhaustive: clippy suggested any(),
  which would stop at the first changed scan.
- Follow the existing per-function allow for too_many_arguments, and
  allow large_enum_variant on long-lived state and low-rate messages
  where boxing would only add allocations.
- Keep one dead_code allow per module instead of duplicated inner and
  outer attributes; launcher_screensaver keeps its own because the
  preview binary mounts it separately.

CI's app group now also runs clippy on the ui library and tests with
warnings denied.


* Keep the statvfs block-size cast needed on 32-bit ARM

The clippy cleanup removed stats.f_frsize as u64 as an unnecessary
cast, which is true on 64-bit hosts but not on armv7, where f_frsize
is u32, so the device build stopped compiling. Restore it with a scoped
allow and a note on why.


* Fix the Linux-only findings from the new ui clippy step

CI runs the ui clippy step on Linux, which reported findings the macOS
run could not see:

- Rewrite the getrusage/clock_gettime .then(..).unwrap_or(0) chains in
  particle_renderer and launcher_scheduler as if/else; that code is
  only compiled on Linux.
- statvfs field widths differ between macOS, 64-bit Linux and 32-bit
  ARM, so whichever cast is a no-op depends on the target. Allow
  unnecessary_cast on filesystem_available_bytes and keep both casts.

Verified with the CI command inside the Linux build container.


---------

+++ apps/mister/src/ui_runner/launcher_present/latch.rs
@@ -768,7 +768,7 @@ impl<B: LatchFrameBuffers> FpgaVblankLatchHiddenPresenter<B> {
         let copied_rows = if self.render_height == self.height {
             plan.restore_rects
                 .iter()
-                .map(|rect| rect.rows() as u32)
+                .map(|rect| rect.rows())
                 .sum::<u32>()
         } else {
             let vertical_transform =
